change image for bspwm, an add sound for notification, no for audio

+++ dunst/scripts/lognotifs.sh
@@ -7,9 +7,11 @@ crunch_icon=$(echo "$4" | sed '/^$/d')
 crunch_urgency=$(echo "$5" | sed '/^$/d')
 timestamp=$(date +"%I:%M %p")
 
-audacious -H /usr/share/sounds/freedesktop/stereo/message.oga &
-sleep 3
-killall audacious
+if [ "$crunch_appname" != "audio" ]; then
+  audacious -H /usr/share/sounds/freedesktop/stereo/message.oga &
+  sleep 3
+  killall audacious
+fi
 
 if [[ "$crunch_appname" == "Spotify" ]]; then
     random_name=$(mktemp --suffix ".png")
